==> 1c. BallLoader2019.ipynb <==
### Purpose of this script:

Extract ball tracking data to accompany player tracking data for analysis and modeling

import boto3
import pandas as pd, numpy as np
import json
import s3fs
import re
from datetime import datetime
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

s3b = 'nyg-hackathon-154843675742'
dsource = '../common/data/'
balldsource = 's3://nyg-hackathon-154843675742/source_file/nfl/balltracking_game_play/year=2019/'
masterplay = 'master_play_table_v1.csv'

s3 = boto3.resource('s3')
fs = s3fs.S3FileSystem()

ballfiles = fs.ls(balldsource)

# Use master play file to figure out names of ball files
plays = pd.read_csv(dsource+masterplay, low_memory=False)
rsgames = list(plays.gameId.astype("str"))

rsballfiles=[k for k in ballfiles if re.search("(?<=play_)[0-9]*", k).group(0) in rsgames]

def loadjson(fname):
    obj = s3.Object(s3b, re.search("source.*",fname).group(0))
    datastr = obj.get()['Body'].read().decode('utf-8')
    datadict = json.loads(datastr)
    return datadict

len(rsgames), len(rsballfiles), len(ballfiles)

balldf=pd.DataFrame(js["ballTrackingData"])[["time", "x", "y","event"]]

masterballdf = pd.DataFrame(np.zeros((len(rsballfiles),2)).astype('int'))
masterballdf.columns = ["gameId", "playId"]

gameId = list()
playId = list()
tracklist = list()

for k, file in enumerate(rsballfiles):
        
    try:
        js = loadjson(file)
        
        balldf=pd.DataFrame(js["ballTrackingData"])[["time", "x", "y","event"]]

        # Presuming the data is in proper time order, with no missing timeslices, we don't actually need to parse time...
        #####
        # balldf['dt']=[datetime.strptime(x[:-1], "%Y-%m-%dT%H:%M:%S.%f") for x in balldf.time]
        # balldf.drop('time', axis=1, inplace=True)
        # balldf['snaprel']=((balldf.dt-balldf.dt.iloc[0]).astype('int')//10e7).astype('int32')

        try:
            snaprow=(balldf.event=='ball_snap').values.nonzero()[0][0]
            
            balldf=balldf.iloc[snaprow:]
            balldf.drop(['event', 'time'], axis=1, inplace=True)

            gameId.append(js['gameId'])
            playId.append(js['gsisPlayId'])
            tracklist.append(balldf.values)
        except:
            next
        
    except:
        print("failed on "+str(k)+" "+file)

masterdf = pd.DataFrame({"gameId": gameId, "playId": playId, "ballTrack": tracklist})

masterdf.info()

masterdf.head()

import pickle

with open('../common/data/balltrack.pkl', 'wb') as f:
    pickle.dump(masterdf, f)
