==> 4. Create_Blitz_Dataset_w_Engineered_Features.ipynb <==
import pandas as pd
import numpy as np

#Pull in game data
out_path = '../common/data/'

#player data
player = pd.read_pickle(out_path + 'player_play_rect.pkl')
#play data
play = pd.read_pickle(out_path + 'play_rect.pkl')
play_df = pd.DataFrame(play, columns=['playId', 'gameId', 'play_type',
                                      'yardlineNumber', 'playDesc', 
                                      'dInBox','dRushers', 'sack', 'qb_hit']) 
#ball position data
ball_pos = pd.read_pickle(out_path + 'ball_rect.pkl')

#identified blitz plays
blitz = pd.read_csv(out_path + 'blitz_plays.csv')

play.shape

player['unique_key'] = player['gameId'].astype(str) + "_"  + player['playId'].astype(str)
play['unique_key'] = play['gameId'].astype(str) + "_"  + play['playId'].astype(str)
ball_pos['unique_key'] = ball_pos['gameId'].astype(str) + "_"  + ball_pos['playId'].astype(str)
blitz['unique_key'] = blitz['gameId'].astype(str) + "_"  + blitz['playId'].astype(str)

blitz_keys = blitz['unique_key'].to_list()

player_blitz_cols = ['unique_key','displayName', 'jerseyNumber', 'teamType', 'position', 'positionGroup', 'coordinates']

play.columns

play_blitz_cols = ['gameId', 'playId', 'snapTime', 'endTime','unique_key','home_team', 'away_team', 'homeScore',
       'visitorScore', 'quarter', 'gameClock', 'possessionTeam', 'playDesc',
       'down', 'ydstogo', 'play_type', 'isGoalToGo', 'isRedzonePlay', 'isScoring',
       'isPenalty', 'oForm', 'dInBox', 'RB', 'TE', 'WR', 'DL', 'LB', 'DB',
       'dRushers', 'shotgun', 'no_huddle', 'qb_dropback', 'qb_scramble',
       'pass_location', 'air_yards', 'run_location', 'run_gap', 'ep', 'epa',
       'wp', 'home_wp', 'wpa', 'first_down_rush', 'first_down_pass',
       'third_down_converted', 'third_down_failed', 'fourth_down_converted',
       'fourth_down_failed', 'incomplete_pass', 'interception',
       'fumble_forced', 'tackled_for_loss', 'fumble_lost', 'qb_hit',
       'rush_attempt', 'pass_attempt', 'sack', 'fumble', 'complete_pass',
       'tackle_for_loss_1_player_id', 'tackle_for_loss_1_player_name',
       'qb_hit_1_player_id', 'qb_hit_1_player_name', 'qb_hit_2_player_id',
       'qb_hit_2_player_name', 'forced_fumble_player_1_player_id',
       'forced_fumble_player_1_player_name', 'rectYardline', 'dir', 'snapPos']

ball_pos_blitz_cols = ['ballTrack', 'unique_key']

player_blitz = player[player['unique_key'].isin(blitz_keys)][player_blitz_cols].copy()
play_blitz = play[play['unique_key'].isin(blitz_keys)][play_blitz_cols].copy()
ball_pos_blitz = ball_pos[ball_pos['unique_key'].isin(blitz_keys)][ball_pos_blitz_cols].copy()

df_merged = player_blitz.merge(play_blitz, on='unique_key', how='left')
df_merged = df_merged.merge(ball_pos_blitz, on='unique_key', how='left')

# play_blitz[['qb_scramble', 'incomplete_pass', 'interception', 'sack', 'complete_pass', 'qb_hit']].sum()

# play_blitz[(play_blitz['complete_pass'] == 1) & (play_blitz['qb_hit'] == 1)]

# play_blitz['blitz_successful'] = play_blitz.apply(lambda row: 1 if (row['qb_scramble'] == 1) |
#                                                                      (row['incomplete_pass'] == 1) |
#                                                                      (row['qb_hit'] == 1) |
#                                                                     (row['interception'] == 1) |
#                                                                     (row['sack'] == 1 ) else 0, axis=1)

# play_blitz.blitz_successful.value_counts()

team_mapping = {'DB': 'defense', 'OL': 'offense', 'LB': 'defense', 'DL': 'defense', 
                'WR': 'offense', 'RB': 'offense', 'TE': 'offense', 'QB': 'offense'}

df_merged['player_side'] = df_merged.positionGroup.map(team_mapping)

#pull xy coordinates for snap & play end
for i in df_merged:
    x_snap = lambda x: x[20][0]
    y_snap = lambda x: x[20][1]
    ball_x_snap = lambda x: x[20][0]
    ball_y_snap = lambda x: x[20][1]
    x_1s = lambda x: x[30][0]
    y_1s = lambda x: x[30][1]
    x_3s = lambda x: x[50][0]
    y_3s = lambda x: x[50][1]
    x_2s = lambda x: x[40][0]
    y_2s = lambda x: x[40][1]
    x_end = lambda x: x[-11][0]
    y_end = lambda x: x[-11][1]
    
    df_merged['x_snap'] = df_merged['coordinates'].apply(x_snap)
    df_merged['y_snap'] = df_merged['coordinates'].apply(y_snap)
    df_merged['x_snap_ball'] = df_merged['ballTrack'].apply(ball_x_snap)
    df_merged['y_snap_ball'] = df_merged['ballTrack'].apply(ball_y_snap)
    df_merged['x_1s'] = df_merged['coordinates'].apply(x_1s)
    df_merged['y_1s'] = df_merged['coordinates'].apply(y_1s)
    df_merged['x_3s'] = df_merged['coordinates'].apply(x_3s)
    df_merged['y_3s'] = df_merged['coordinates'].apply(y_3s)
    df_merged['x_2s'] = df_merged['coordinates'].apply(x_2s)
    df_merged['y_2s'] = df_merged['coordinates'].apply(y_2s)
    df_merged['x_end'] = df_merged['coordinates'].apply(x_end)
    df_merged['y_end'] = df_merged['coordinates'].apply(y_end)
    

    
# print(df_merged.info())

#add qb locations to all rows in play
ds_qb = df_merged[df_merged['position'] == 'QB'].copy()
ds_qb['qb_x_snap'] = ds_qb['x_snap']
ds_qb['qb_x_1s'] = ds_qb['x_1s']
ds_qb['qb_x_2s'] = ds_qb['x_2s']
ds_qb['qb_x_3s'] = ds_qb['x_3s']
ds_qb['qb_y_snap'] = ds_qb['y_snap']
ds_qb['qb_y_1s'] = ds_qb['y_1s']
ds_qb['qb_y_2s'] = ds_qb['y_2s']
ds_qb['qb_y_3s'] = ds_qb['y_3s']
ds_qb = ds_qb[['unique_key', 'qb_x_snap', 'qb_x_1s', 'qb_x_2s', 'qb_x_3s', 'qb_y_snap', 'qb_y_1s', 'qb_y_2s', 'qb_y_3s']]
df_merged = pd.merge(df_merged, ds_qb, how = 'left', on = 'unique_key')

df_merged['x_player_pos_v_LOS'] = df_merged['x_snap'] - df_merged['x_snap_ball']
df_merged['y_player_pos_v_LOS'] = df_merged['y_snap'] - df_merged['y_snap_ball']
df_merged['x_1s_toLOS'] = df_merged['x_1s'] - df_merged['x_snap_ball']
df_merged['y_1s_toLOS'] = df_merged['y_1s'] - df_merged['y_snap_ball']
df_merged['x_2s_toLOS'] = df_merged['x_2s'] - df_merged['x_snap_ball']
df_merged['y_2s_toLOS'] = df_merged['y_2s'] - df_merged['y_snap_ball']
df_merged['x_3s_toLOS'] = df_merged['x_3s'] - df_merged['x_snap_ball']
df_merged['y_3s_toLOS'] = df_merged['y_3s'] - df_merged['y_snap_ball']
df_merged['qb_dist_snap'] = ((df_merged['x_snap'] - df_merged['qb_x_snap'])**2 + (df_merged['y_snap'] - df_merged['qb_y_snap'])**2)**0.5
df_merged['qb_dist_1s'] = ((df_merged['x_1s'] - df_merged['qb_x_1s'])**2 + (df_merged['y_1s'] - df_merged['qb_y_1s'])**2)**0.5
df_merged['qb_dist_2s'] = ((df_merged['x_2s'] - df_merged['qb_x_2s'])**2 + (df_merged['y_2s'] - df_merged['qb_y_2s'])**2)**0.5
df_merged['qb_dist_3s'] = ((df_merged['x_3s'] - df_merged['qb_x_3s'])**2 + (df_merged['y_3s'] - df_merged['qb_y_3s'])**2)**0.5

df_merged.to_pickle('../common/data/blitz_player_play.pkl')
