==> 5. Blitz_Plays_Analysis.ipynb <==
import boto3
import pandas as pd
import numpy as np
import math
from sagemaker import get_execution_role
import json
import s3fs
import gzip
import pickle

#Pull in game data
out_path = '../common/data/'

#blitz plays
blitz = pd.read_csv(out_path + 'blitz_plays.csv')
#player data
player = pd.read_pickle(out_path + 'player_play_rect.pkl')
#play data
play = pd.read_pickle(out_path + 'play_rect.pkl')
#play_df = pd.DataFrame(play, columns=['playId', 'gameId', 'play_type',
#                                      'yardlineNumber', 'playDesc', 
#                                      'dInBox','dRushers', 'sack', 'interception', 'qb_hit']) 

#Merge datasets against blitz plays
blitz_d = pd.merge(blitz, play, how='left', on=['playId', 'gameId'])
blitz_df = pd.merge(blitz_d, player, how='left', on=['playId', 'gameId'])
team_mapping = {'DB': 'defense', 'OL': 'offense', 'LB': 'defense', 'DL': 'defense', 
                'WR': 'offense', 'RB': 'offense', 'TE': 'offense', 'QB': 'offense'}
blitz_df['player_side'] = blitz_df.positionGroup.map(team_mapping)
blitz_df = blitz_df[(blitz_df['player_side'] == 'defense') | (blitz_df.position == 'QB')] #limit to defensive players & QBs

blitz['playId'].nunique

blitz_df['playId'].nunique #check merge

pd.set_option('display.max_columns', None)
blitz_df.head(5)

# Develop Features

#pull xy coordinates for important points in the play - player, LOS
x=0
for i in blitz_df:
    x_snap = lambda x: x[20][0]
    y_snap = lambda x: x[20][1]
    LOS_x = lambda x: x[0]
    LOS_y = lambda x: x[1]
    x_1s = lambda x: x[30][0]
    y_1s = lambda x: x[30][1]
    x_2s = lambda x: x[40][0]
    y_2s = lambda x: x[40][1]
    x_3s = lambda x: x[50][0]
    y_3s = lambda x: x[50][1]
    x_end = lambda x: x[-11][0]
    y_end = lambda x: x[-11][1]
    
    blitz_df['x_snap'] = blitz_df['coordinates'].apply(x_snap)
    blitz_df['y_snap'] = blitz_df['coordinates'].apply(y_snap)
    blitz_df['LOS_x'] = blitz_df['snapPos_x'].apply(LOS_x)
    blitz_df['LOS_y'] = blitz_df['snapPos_x'].apply(LOS_y)
    blitz_df['x_1s'] = blitz_df['coordinates'].apply(x_1s)
    blitz_df['y_1s'] = blitz_df['coordinates'].apply(y_1s)
    blitz_df['x_2s'] = blitz_df['coordinates'].apply(x_2s)
    blitz_df['y_2s'] = blitz_df['coordinates'].apply(y_2s)
    blitz_df['x_3s'] = blitz_df['coordinates'].apply(x_3s)
    blitz_df['y_3s'] = blitz_df['coordinates'].apply(y_3s)
    blitz_df['x_end'] = blitz_df['coordinates'].apply(x_end)
    blitz_df['y_end'] = blitz_df['coordinates'].apply(y_end)
    
    
print(blitz_df.info())

#Assess pass rushers
blitz_df['x_player_pos_v_LOS'] = blitz_df['x_snap'] - blitz_df['LOS_x']
blitz_df['y_player_pos_v_LOS'] = blitz_df['y_snap'] - blitz_df['LOS_y']
blitz_df['x_1s_toLOS'] = blitz_df['x_1s'] - blitz_df['LOS_x']
blitz_df['y_1s_toLOS'] = blitz_df['y_1s'] - blitz_df['LOS_y']
blitz_df['x_2s_toLOS'] = blitz_df['x_2s'] - blitz_df['LOS_x']
blitz_df['y_2s_toLOS'] = blitz_df['y_2s'] - blitz_df['LOS_y']
blitz_df['x_3s_toLOS'] = blitz_df['x_3s'] - blitz_df['LOS_x']
blitz_df['y_3s_toLOS'] = blitz_df['y_3s'] - blitz_df['LOS_y']
#Define pass rushers
blitz_df['crossLOS']  = (blitz_df['x_2s_toLOS'] > -0.2)
blitz_df['crossLOS2']  = (blitz_df['x_1s_toLOS'] > -0.2)
blitz_df['y_bounds'] = ((abs(blitz_df['y_1s'] - blitz_df['LOS_y']) < 7))
blitz_df['y_bounds2'] = ((abs(blitz_df['y_3s'] - blitz_df['LOS_y']) < 6))
blitz_df['rusher'] = (((blitz_df['crossLOS2'] == True) | (blitz_df['crossLOS'] == True)) & ((blitz_df['y_bounds'] == True) | (blitz_df['y_bounds2'] == True))) 

#Separate out QBs & load back in
qb = blitz_df[blitz_df.position == 'QB']
qb['qb_xend'] = qb['x_end'] #QB at end of play
qb['qb_yend'] = qb['y_end'] #QB at end of play
qb = pd.DataFrame(qb, columns=['playId', 'gameId', 'qb_xend', 'qb_yend'])
blitz_df = pd.merge(blitz_df, qb, how='left', on=['playId', 'gameId'])

blitz_df['playId'].nunique #check merge

#Create features 
blitz_df['vertPlay'] = abs(blitz_df['x_end'] - blitz_df['x_snap'])
blitz_df['horizPlay'] = abs(blitz_df['y_end'] - blitz_df['y_snap'])
blitz_df['vert1s'] = abs(blitz_df['x_1s'] - blitz_df['x_snap'])
blitz_df['vert3s'] = abs(blitz_df['x_3s'] - blitz_df['x_snap'])
blitz_df['horiz3s'] = abs(blitz_df['y_3s'] - blitz_df['y_snap'])
blitz_df['xfromQB'] = abs(blitz_df['qb_xend'] - blitz_df['x_end'])
blitz_df['yfromQB'] = abs(blitz_df['qb_yend'] - blitz_df['y_end'])
blitz_df['QBHurry'] = blitz_df['xfromQB'] < 1
blitz_df['vertSpeed_1s'] = blitz_df['vert1s']/3
blitz_df['vertSpeed_3s'] = blitz_df['vert3s']/3
#blitz_df['magnitude'] = math.sqrt((blitz_df['y_end']-blitz_df['y_snap'])**2 + (blitz_df['x_end']-blitz_df['x_snap'])**2)
#blitz_df['angle'] = math.atan2((blitz_df['y_end']-blitz_df['y_snap']),(blitz_df['x_end']-blitz_df['x_snap']))*(180/(math.pi))

# Analyze Features

#Pull in game data
out_path = '../common/data/'
blitz_df = pd.read_pickle(out_path + 'blitz_player_play.pkl')
blitz_df = blitz_df[(blitz_df['player_side'] == 'defense')] #limit to defensive players
blitz_df['unique_player_key'] = blitz_df['displayName'].astype(str) + "_"  + blitz_df['jerseyNumber'].astype(str) #add unique player ID
#blitz_df.head(5)

#Add in rusher methodology
blitz_df['crossLOS']  = (blitz_df['x_2s_toLOS'] > -0.2)
blitz_df['crossLOS2']  = (blitz_df['x_1s_toLOS'] > -0.2)
blitz_df['y_bounds'] = ((abs(blitz_df['y_1s'] - blitz_df['y_snap_ball']) < 7))
blitz_df['y_bounds2'] = ((abs(blitz_df['y_3s'] - blitz_df['y_snap_ball']) < 6))
blitz_df['rusher'] = ((blitz_df['player_side'] == 'defense') & (((blitz_df['crossLOS2'] == True) | (blitz_df['crossLOS'] == True)) & ((blitz_df['y_bounds'] == True) | (blitz_df['y_bounds2'] == True)))) 

#blitz_df.to_pickle(out_path + 'blitz_player_play.pkl')

#Create features 
blitz_df['vertPlay'] = abs(blitz_df['x_end'] - blitz_df['x_snap'])
blitz_df['horizPlay'] = abs(blitz_df['y_end'] - blitz_df['y_snap'])
blitz_df['vert1s'] = abs(blitz_df['x_1s'] - blitz_df['x_snap'])
blitz_df['vert3s'] = abs(blitz_df['x_3s'] - blitz_df['x_snap'])
blitz_df['horiz3s'] = abs(blitz_df['y_3s'] - blitz_df['y_snap'])
blitz_df['xfromQB'] = abs(blitz_df['qb_xend'] - blitz_df['x_end'])
blitz_df['yfromQB'] = abs(blitz_df['qb_yend'] - blitz_df['y_end'])
blitz_df['QBHurry'] = blitz_df['xfromQB'] < 1
blitz_df['vertSpeed_1s'] = blitz_df['vert1s']/3
blitz_df['vertSpeed_3s'] = blitz_df['vert3s']/3

#Define success metric
blitz_df['success'] = (blitz_df['sack'] == 1) | (blitz_df['interception'] == 1)

#################################################################
###              Create rusher stats dataset                  ###
#################################################################

rusher = blitz_df[(blitz_df['rusher'] == True)]

#Who consistently gets closest to the QB
xQB = rusher.groupby(['displayName', 'position', 'unique_player_key'])['xfromQB'].mean().reset_index(name='avg_xfromQB')
xQB.sort_values(by='avg_xfromQB', ascending=True)

#Who is getting closest to the QB the most
QBh = rusher[rusher['QBHurry'] == True].groupby(['displayName', 'position', 'unique_player_key'])['QBHurry'].count().reset_index(name='QBHurry_cnt')
QBh.sort_values(by='QBHurry_cnt', ascending=False)

#Who is the fastest in 1 sec
s_1 = rusher.groupby(['displayName', 'position', 'unique_player_key'])['vertSpeed_1s'].max().reset_index(name='max_vertSpeed_1s')
s_1.sort_values(by='max_vertSpeed_1s', ascending=False)

#Who is the fastest in 3 sec - rusher
s_3 = rusher.groupby(['displayName', 'position', 'unique_player_key'])['vertSpeed_3s'].max().reset_index(name='max_vertSpeed_3s')
s_3.sort_values(by='max_vertSpeed_3s', ascending=False)

#Who is consistently fast in 1 sec - rushers
sa_1 = rusher.groupby(['displayName', 'position', 'unique_player_key'])['vertSpeed_1s'].mean().reset_index(name='avg_vertSpeed_1s')
sa_1.sort_values(by='avg_vertSpeed_1s', ascending=False)

#Who is consistently fast in 3 sec - rushers
sa_3 = rusher.groupby(['displayName', 'position', 'unique_player_key'])['vertSpeed_3s'].mean().reset_index(name='avg_vertSpeed_3s')
sa_3.sort_values(by='avg_vertSpeed_3s', ascending=False)

#Who is rushing the most
vol = rusher.groupby(['displayName', 'position', 'unique_player_key'])['playId'].count().reset_index(name='rush_cnt')
vol.sort_values(by='rush_cnt', ascending=False)

#Who is causes the most success
suc = rusher[rusher['success'] == True].groupby(['displayName', 'position', 'unique_player_key'])['success'].count().reset_index(name='success_cnt')
suc.sort_values(by='success_cnt', ascending=False)

#Export rusher dataset
rusher_exp = rusher['unique_player_key'].drop_duplicates()
#rusher_exp.shape
rusher_exp = pd.merge(rusher_exp, xQB[['unique_player_key','position', 'avg_xfromQB']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, QBh[['unique_player_key', 'QBHurry_cnt']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, s_1[['unique_player_key', 'max_vertSpeed_1s']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, s_3[['unique_player_key', 'max_vertSpeed_3s']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, sa_1[['unique_player_key', 'avg_vertSpeed_1s']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, sa_3[['unique_player_key', 'avg_vertSpeed_3s']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, vol[['unique_player_key', 'rush_cnt']], how='left', on=['unique_player_key'])
rusher_exp = pd.merge(rusher_exp, suc[['unique_player_key', 'success_cnt']], how='left', on=['unique_player_key'])
#rusher_exp.head(10)
#rusher_exp.shape
rusher_exp.to_csv(out_path + 'rusher_metrics.csv')

#rusher_exp[rusher_exp['unique_player_key'] == 'Khalil Mack_52']
